File-Sharing-Bot/requirements.txt
python-telegram-bot==10.1.0
